requirements.txt
atlassian-python-api
beautifulsoup4
markdownify
ragflow-sdk
requests